Add explicit size_t conversions.

+++ src/libsodium/crypto_pwhash/scryptsalsa208sha256/pwhash_scryptsalsa208sha256.c
@@ -34,7 +34,7 @@ pickparams(unsigned long long opslimit, const size_t memlimit,
             }
         }
     } else {
-        maxN = memlimit / (*r * 128);
+        maxN = memlimit / ((size_t) *r * 128);
         for (*N_log2 = 1; *N_log2 < 63; *N_log2 += 1) {
             if ((uint64_t) (1) << *N_log2 > maxN / 2) {
                 break;
